+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 3.2.8 (2025-07-01)
 
 * (theimo1221) Add AC-Battery setting, to allow cooling sooner on sunny days even though the battery is kinda empty.